backend/requirements.txt
fastapi
uvicorn
langchain
langchain_community
chromadb
sentence-transformers
ollama
google-auth
google-api-python-client
slack-sdk
python-dotenv
